add displaying of updated list
