+++ package.json
@@ -17,7 +17,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "styled-components": "^5.3.8",
-    "swr": "^2.1.1"
+    "swr": "^2.1.1",
+    "uid": "^2.0.2"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
