+++ pages/players/index.js
@@ -1,33 +1,5 @@
-import useSWR from "swr";
-import styled from "styled-components";
+import List from "../../components/List";
 
-const fetcher = (...args) => fetch(...args).then((res) => res.json());
-
-export default function Players() {
-  const { data, isLoading } = useSWR("/api/players", fetcher, {
-    fallbackData: [],
-  });
-
-  if (isLoading) return <div>loading...</div>;
-
-  if (data.length === 0) {
-    return <div>Currently no players are saved</div>;
-  }
-
-  return (
-    <StyledList>
-      {data.map((player) => {
-        return (
-          <li key={player.response[0].player.id}>
-            {player.response[0].player.name}
-          </li>
-        );
-      })}
-    </StyledList>
-  );
+export default function Players({ players, setPlayers }) {
+  return <List players={players} />;
 }
-
-const StyledList = styled.ul`
-  list-style-type: none;
-  margin: 20% 20%;
-`;
